+++ src/examples/t003.go
@@ -15,8 +15,8 @@ func main() {
 	}
 	defer p.Close()
 
-	p.PlotXY([]float{0,1,2,3,4,5,6,7,8,9,10}, 
-		     []float{0,1,2,3,4,5,6,7,8,9,10}, "some data - x/y")
+	p.PlotXY([]float64{0,1,2,3,4,5,6,7,8,9,10}, 
+		     []float64{0,1,2,3,4,5,6,7,8,9,10}, "some data - x/y")
 	p.SetXLabel("my x data")
 	p.SetYLabel("my y data")
 	p.CheckedCmd("set terminal pdf")
